pyproject.toml
[project]
name = "magi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Shawn Eddy", email = "ymiseddy@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=1.67.0",
    "pydantic-ai-skills>=0.5.1",
    "pydotenv>=0.0.7",
    "pyyaml>=6.0.3",
    "typing-extensions>=4.15.0",
]

[project.scripts]
magi = "magi:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
